Separate GuildWelcomeScreen.Channel impl
